+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_params.py
@@ -435,8 +435,8 @@ def load_arguments(self, _):
                        help="storage caching type for data disk(s), including 'None', 'ReadOnly', 'ReadWrite', etc. Use a singular value to apply on all disks, or use '<lun>=<vaule1> <lun>=<value2>' to configure individual disk")
             c.argument('ultra_ssd_enabled', arg_type=get_three_state_flag(), min_api='2018-06-01',
                        help='(PREVIEW) Enables or disables the capability to have 1 or more managed data disks with UltraSSD_LRS storage account')
-            c.argument('use_local_os_disk', arg_type=get_three_state_flag(), min_api='2018-06-01',
-                       help='(Preview) Use a managed ephemeral (local) os disk. For more info, see here: https://azure.microsoft.com/en-us/blog/ephemeral-os-disk-limited-public-preview/')
+            c.argument('ephemeral_os_disk', arg_type=get_three_state_flag(), min_api='2018-06-01',
+                       help='(Preview) Allows you to create an OS disk directly on the host node, providing local disk performance and faster VM/VMSS reimage time.')
 
         with self.argument_context(scope, arg_group='Network') as c:
             c.argument('vnet_name', help='Name of the virtual network when creating a new one or referencing an existing one.')